# docs/openapi.yaml
openapi: 3.0.3
info:
  title: cardinal-wings API
  description: >
    REST API daemon for managing cardinal over HTTP. Panel-oriented surface
    over cardinal's Docker-compatible API, with multi-node routing (`?node=`),
    async tasks, streaming logs/exec, metrics and role-based auth.
  version: "0.2.0-dev"
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/containers/{id}/backup:
    get:
      tags: [containers]
      summary: Download container data as tar.gz (live stream)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { "200": { description: tar.gz archive } }
    post:
      tags: [containers]
      summary: Snapshot ({"uuid"} → 202 async), stored restore ({"backup_uuid"}/{ "download_url"}), or binary tar.gz upload; ?clean=1 for exact restore (admin)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                uuid: { type: string }
                backup_uuid: { type: string }
                download_url: { type: string }
                format: { type: string }
      responses: { "200": { description: restored }, "202": { description: snapshot accepted } }
    delete:
      tags: [containers]
      summary: Delete a stored snapshot (?uuid=, admin; 404 when already gone)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: uuid, in: query, schema: { type: string } }
      responses: { "200": { description: deleted } }
...
  /download/backup:
    get:
      tags: [system]
      summary: Stored snapshot download via panel-signed JWT (?token=, scope backup-download)
      parameters: [{ name: token, in: query, required: true, schema: { type: string } }]
      responses: { "200": { description: tar.gz snapshot } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...